+++ src/main/java/chess/domain/Piece.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain;
 
 import java.util.HashMap;
 import java.util.Map;
