+++ src/main/java/chess/domain/PieceType.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain;
 
 public enum PieceType {
