+++ src/main/java/chess/domain/Team.java
@@ -1,7 +1,7 @@
-package chess;
+package chess.domain;
 
 public enum Team {
-    
+
     BLACK,
     WHITE,
     ;
